Change MySQL port from 3306 to 5678
This is to prevent interference with other services that might be
running on port 3306 on the host computer, since 3306 is a popular port
selection, especially among database applications.

No specific reason for port 5678, it's just easy to type. We might want
to choose something else, since this could easily interfere with a
developer that uses this number for something else (it's easy to type),
but it's better than 3306.

Fixes #77.