Use fresh data for transactions (#139)
* Add reducer with some basic actions

The reducer does define some actions, but never produces a different
state.

* Add function to initalize the data store for fresh-data

* Add function to initialize the fresh-data API client

* Add empty Payments API spec

This will be filled in as we go, but can remain empty for now since
we're not actually calling any Payments endpoints for data.

* Add dependencies to package.json

* Add function to initialize the Payments data store API

* Include the fresh-data Payments data store in web app

* Refactor TransactionsList to React Component class

This is necessary to use the component with fresh-data. Currently the
component always shows an empty table.

* Add withSelect function

This is used to wrap components before exporting them, and allows
components to access the fresh-data data store.

* Wrap TransactionsList export with `withSelect`

This will allow us to access the fresh-data data store.

* Add hard coded test data to transactions fresh-data selector

The selector really shouldn't be using hard coded data, but this is for
testing! All for the greater good, I'm sure.

* Add transactions selector to the Payments API spec

* Retrieve data to display in TransactionsList using transactions selector

By retrieving the relevant selectors we can access the data from the
fresh-data data store, and use that to populate the transactions table.

As far as I understand this (@coderkevin would be a better person to
ask) the `withSelect` method basically injects data from the selector
into the component's `this.props`, allowing us to use the data in the
component.

* Update package-lock.json

* Fix wrong parameter name in `createApiClient`

* Add reducer for `fresh-data`

* Add lodash to dependencies

* Add read operation for transactions list

This operation defines how requests are sent for `fresh-data`
transactions resources, i.e. how the transactions list is retrieved.

* Expose the transactions list 'read' operation to `fresh-data`

* Add transactions list selector

The selector can be used to retrieve transactions list for components.
This commit also adapts the code to fit the new selector better.

* Add code to determine if the table loading view should be displayed

* Fix `getTransactionsLoading` function

Incorrectly accessing the resource object caused the timestamps to be
undefined when evaluating whether the transactions request is loading or
not.

* Update fresh-data version in package.json

* Change loading status selector to not specify requirements

* Export transactions operations to improve testability

* Remove unnecessary setup code from fresh-data data store initialization

* Add packages necessary to run tests

For some reason jest can't find '@wordpress/api-fetch' unless it's
installed. Even though the test doesn't actually _use_ api-fetch, it
complains because one of the included files imports api-fetch.

* Fix transactions operations

* Add transactions operation tests

* Add transactions selector tests

* Merge changes to transactions-list.js from master

* Add `showTransactionsPlaceholder` selector

Used to indicate whether the `TableCard` in the transactions list should
show the loading placeholder view or not.

* Fix `...` (spread operator) missing error.

Babel seems to need a specific package to support the `...` operator for
now.
See: https://github.com/babel/babel/issues/10179#issuecomment-510254948

npm i --save-dev @babel/plugin-proposal-object-rest-spread is sufficient
as a fix.