Use the transactions cache on wpcom to populate transactions list (#341)
* Fix transaction list

After flattening the data by using a cache on the server side the
transactions table no longer populated. This commit fixes that, although
we lost the charge ID, so Order Information doesn't fully populate.

* Add transactions summary endpoint to local site API

Adds the /transactions/summary endpoint to the API to allow populate the
summary for the transactions table.

* Add summary to transactions table

Adds the summary at the bottom of the transactions table. While the
summary is loading nothing is displayed, but once the loading is done,
the summary is added.

* Update tests to match new data format

* Add transactions reducer tests

* Make transaction summary selectors a bit more robust

* Add tests for transactions selectors

* Update client code to consume updated data format from server

This commit updates the client API and front-end code to fully accept
the "finalized" version of the transactions server endpoint. We now have
the previously missing information and can properly populate the
transactions table.

* Re-introduce moment so correct time is displayed for different timezones

* Add a RiskLevel component

Contains logic for how to display risk level.

* Fix indentation in file unrelated to this feature

* Remove the `get_charge_id_from_transaction` function

It now sort of just complicates the code since with the latest changes
server side we can now just get the charge id like so:
`$transaction['charge_id']`.