Handle failed refunds (#411)
* Extract DB access from Stripe API client class

This logic is required by other parts of the code base, so extracting
it out allows it to be shared. The unit tests have been left as is for
now, but this also allows for database access to be mocked more easily.

* Add REST endpoint for handling webhooks from WP.com

The initial implementation just handles refund updated events and adds
an order note in the event of a failure.

* Turn off Squiz.Commenting.FunctionCommentThrowTag.WrongNumber PHPCS rule

This rule is generating some false positives. We can try turning it on
again if it gets fixed in a later PHPCS version. Alternatively we could
start relying on a static code analyser for more robust checks on how
we're using exceptions.