Add 3DS support for pay for order page (#745)
* Checkout: Move auth modal logic to a function

* Show authentication modal on Pay for Order page

This change calls `maybeShowAuthenticationModal()`
on every page load. The function will check if the modal
needs to be displayed before displaying it.

* Checkout: Move form blocking logic to function

* On Pay for Order block payment UI during auth

On the Pay for Order page, authentication
happens after a page reload. This causes the payment form
to lose the credit card info previously entered. To avoid user
confusion, this form should be hidden and blocked from
being submitted.

After the payment is complete, the user is redirected. If there
was an error processing the authentication, the payment form is
unblocked.

* Verify intent used for auth same as order intent

Send the intent ID along with the request to
update the order status from the JS script,
so that the intent can be matched with the one
we already have saved for the order.

Introduce a new custom exception class,
WC_Payments_Authentication_Intent_Mismatch_Exception,
to deal with mismatched intents when handling
payment authentication.

* Remove duplicate variable assignment

* Amend auth exception to handle multiple cases

* Ensure intent used for authentication not empty

* Add comment to say function is used in two cases

* For Pay for Order use order ID in url not hash

Instead of relying on the order ID from the
hash, use the order ID present in the url
pathname for Pay for Order pages.

* Get order ID for all permalinks using full URL

Sometimes the order ID is not found in the `path`,
so the full URL needs to be searched for the order ID.
It can be found after the string `order-pay`.