Subscriptions integration step 1 (#826)
* Fix the token service class so it uses the provided user object instead of the current user

* Added a "Subscription Compat" gateway sub-class to encapsulate all the subscription-related logic

* Implement subscription sign-up and automatic renewal payment

* Ensure Subscriptions are v3.0.0 or newer

* Add class to hold payment information

WC_Payment_Information is intended to hold payment information during
payment processing. It has some utility functions that describe what it
is you're actually checking better than the previously used array of
parameters.

* Adding newly created tokens to subscriptions.

Before this was done, purchasing a subscription with a new card lead to failed renewals since renewals require information from the parent order.

* Remove copy of payment methods from the parent order

Renewal orders were created with the payment_tokens from the parent order, so when changing
the token for a subscription, we'd need to also update the parent order with the new method.

Subscriptions already copies the data from the subscription order when creating a new one,
so we don't need this method unless we really want to use the parent order payment methods.

(cherry picked from commit 87d39ef87421158f158b309a1e827142776cdf2d)

* Fix subscriptions with no payment token

After the copy from the parent order was removed, new subscriptions were not able to
be renewed, because they did not have a payment_token added to it during checkout.
This commit checks if the order has any subscriptions and updates them with the new
payment token.

(cherry picked from commit 249765c1f40d8bce9e77f313532764bfbd6be4ff)

* Add types to function parameters and return values

Mostly an experiment. Made sure to keep types declarations such that
they're supported in PHP 7.0 (our minimum required PHP version), i.e.
there is no usage of `object` or `iterable` types.

See PHP documentation [0] for more information.

[0] https://www.php.net/manual/en/functions.arguments.php#functions.arguments.type-declaration

* Add `add_token_to_order` to the gateway classes

For a subscription `add_token_to_order` adds the provided token to the
provided order and the subscriptions associated with that order.

For simple payments (i.e. non-subscriptions) `add_token_to_order` adds
the provided token to the provided order.

* Change is_recurring to force_save_payment_method

`is_recurring` seemed to be more appropriate as
  1. A parameter to the `process_payment` functions; and
  2. An indicator of whether we're force saving a payment method or not

Because of that we've renamed the parameter to
`force_save_payment_method` and moved it out of the
`Payment_Information` class.

* Derive `is_using_saved_payment_method` from token

Instead of storing an explicit member variable that indicates whether
the payment used a saved payment method, make that decision based on
whether the `Payment_Information` object has a token or not, since
tokens only exist for saved payment methods.

This commit also moves a lot of the complexity of checking for errors
out of the static functions and into the constructor. That way we're
only throwing an exception when trying to instantiate a
`Payment_Information` object with invalid data.

Static functions now return an expected value or `null`.

* Add reminder to add tokens to subs during SCA flow

Should be handled in #833

* Use Payment_Information class in create_setup_intent

After the introduction of the Payment_Information class, the only usage of the
gateway get_payment_method_from_request was in create_setup_intent. Instead of
keeping similar methods, this commit uses the Payment_Information class to
grab the payment method in create_setup_intent and removes the unused
get_payment_method_from_request and get_token_from_request gateway methods.

* Fix unit tests

WC_Payment_Gateway_WCPay_Process_Payment_Test::test_exception_thrown had to be
modified because the notice and status update for failing orders have been moved
to process_payment, which cannot accept a mock order. This commit changes the
mock_order to use WC_Helper_Order::create_order, which is a real WC order and
is saved to the database.

Co-authored-by: Kristófer R <kristofer.thorlaksson@automattic.com>
Co-authored-by: Radoslav Georgiev <rageorgiev@gmail.com>
Co-authored-by: Luiz Reis <luiz.reis@automattic.com>