Add subscriptions unit tests (#875)
* Add Payment_Information tests

* Add token request key variable to get_token_from_request

* Add subscriptions sign up tests

To avoid coupling the tests with WC Subscriptions code, this commit introduces
a WCS_Mock helper that can be used to mock WCS global functions. It also uses
regular orders, which should behave the same as a subscription one once WCS
methods are mocked.

* Add add_token_to_order tests

* Add update_failing_payment_method tests

* Use gateway add_token_to_order when updating failed payment methods

* Add scheduled_subscription_payment tests

* Use current year + 1 as the test tokens expiry date

The test tokens are never really used, so we shoudln't have any issues if the
token expiry date has passed. However, if WC starts validating the expiry date
before creating an already expired token, they could start failing. This commit
uses the next year for the token expiration, to avoid this if it ever happens.