Fix save payment method checkbox for Subscriptions payment method update (#905)
* Hide the "save payment" checkbox during subscription payment method change

The checkbox should not be displayed when dealing with subscriptions payment
method changes, as it's already expected that the card should be saved if it
is a new one, before being attached to that subscription.

* Save payment method if updating a subscription with a new card

Any new payment method that's added to a subscription should be saved. This
commit does that by setting the force_save_payment_method flag when processing
the payment for a subscription payment method change.