Introduced payments-specific exceptions and migrated from using the generic one (#961)
Introduced payments-specific exceptions and migrated from using the generic one

While usage of generic exceptions is not the best practice, it was also suggested to not use them in one of the reviews. Here we introduce several new exceptions and migrating the existing codebase to use it:
- WC_Payments_AddPaymentMethod_Exception (payment creation behaviour)
- WC_Payments_InvalidPaymentMethod_Exception (payment method identification and usage behaviour)
- WC_Payments_ProcessPayment_Exception (payment processing behaviour)

This enables us to provide more sophisticated orders handling (because you can catch specific exceptions and also handle the human-friendly error codes).