Decouple test and live customers (#972)
This PR decouples test and live customers, instead of sharing the same
meta option for both of them. The migration strategy applied here is to
migrate the legacy key to the new live one when account->is_live is true
and to the new test one when it's false.

If an account is not live, we can safely assume that no live mode customers
could be created and avoid losing the test data as part of the optimistic
migration strategy. If a live account has test mode customers, they'll be
discarded when the checkout page is loaded or a card is added.

* Store test and live mode customer IDs in separate options

When a merchant toggles test mode on or off, the current customer saved
for a user is invalidated. This leads to that customer being recreated,
thus losing their saved methods and associated data on the platform.

This commit takes test mode into account when saving the customer ID
option, so a store can have both of those for the same user, preventing
errors when fetching a live customer in test mode and vice-versa.

* Migrate old customer meta to live key

This commit modifies get_customer_id_by_user_id to update the new
customer ID live option, when in live mode if a customer ID is
found in the deprecated key.

* Add changelog entry

* Add get_is_live method to WC_Payments_Account

This method is useful for deciding if the customer data should be
migrated to the live or test key, given that non-live accounts can
only have test mode objects.

* Use account is_live instead of test_mode when migrating customers

We can safely assume that an account that's not live will only have
test customers. So, we don't have to take the live mode optimistic
approach every time. This avoids the loss of test customer data for
test accounts. The optimistic live mode approach is still used when
the account is_live.