Fix order and subscription numbers on transaction list report (#1556)
* Fix order.value when generating the transaction list

The order.value was being incorrectly assigned. Since it was not
used outside of the CSV download functionality we didn't notice
it until the table export was implemented. This commit fixes it,
fixing the report download as well.

* Fix subscriptions.value when generating the transaction list

The subscriptions.value was being incorrectly assigned. Since it
was not used outside of the CSV download functionality we didn't
notice it until the table export was implemented. This commit
fixes it, fixing the report download as well.