Add the store's currency to the list of available currencies (#1917)
The store currency is the currency in which merchants specify all prices. This
means that its exchange rate should be 1.0. E.g. the conversion from USD to USD
should use a rate of 1 since they're the same currency. Because of that, we can
prevent the "Undefined index" error that happens when the store currency is not
part of the "available_currencies" array by manually injecting it there.