Deposits information block on account overview page (#1892)
* Add TypeScript preset for babel to recognize TypeScript syntax

* Update TypeScript settings and split babel & ts-loader tests

It looks like ts-loader and babel in one test set have a collision and
ts-loader starts reporting weird TypeScript errors like "Property has an
implicit any type".

* Fix typo in page set-up

* Update packages

* Bootstrap deposits information component

* Include Deposits Information on Account Overview

* Bootstrap markup for deposits information block

* Disable CSS class name linter checks for TypeScript too

* Simplify set-up in favor of `ts-loader` and remove babel preset.

There are two options in linting TypeScript files. Since we already
have babel and ts-loader running simultaneously it is possible to remove
typescript preset for babel and reduce number of configurations to keep
an eye on.

* Add the `deposits/overview-all` REST route to the client.

* Add all deposits' overviews to the state:

- Action types
- Actions for updating the state.
- The reducer cases to handle those overviews.
- Simple selectors (more logic will be added)
- A hook, which returns the complete deposits' overviews state.

* Update the `getAllDepositsOverviews` selector, as well as all associated type declarations.

* Export functions from the old deposits overview

This is done in order to avoid DRY. Should we move them somewhere else?

* Add the details to the view.

* Adding proper styles to the block.

* Using an integer size for icons.

They're also a bit bigger now, because `20` is not one of the recommended sizes. It was either 18 or 24, and 24 looks much better than 18, which is too small.

* Addressing styling feedback.
- Separate the style import from the rest of the imports.
- Fix the padding of headers by switching their size to small.

* New logic for displaying default currencies first

Using array sorting for this turned out surprisingly unreliable. There is no consistency between Firefox and Chrome.

* Adding instant deposits to the mix

Format the REST API respose accordingly and add the needed components.

NOTE: As of this commit, the neccessary functionality is not in the server yet!

* Adding reponsive styles to card headers

* Extracting deposit overview utilities

Now moved to a separate `client/deposits/utils/index.js` file. it's not top-level placement, but the path makes a lot of sense to me.

* Fixing unut tests with TS

During tests, TypeScript did not like the way everything is exported from the data module, like reducers, etc.

@reykjalin provided this fix, which not only makes the tests work again, but is a code improvement: Every time we import something directly from data it's always a useX hook, so if we just export the hooks that eliminates the issue.

* Isolating `getDepositScheduleDescriptor` in utils.

Those tests used the whole `<DepositsOverview>` component to unit-test a specific function. Now that the function is used in two places, it makes sense to test it directly.

A schedule descriptor is still included in `<DepositsOverview />`, so it's test will make sure that it's included/rendered.

Note: Snapshots were not used for this, hence none are included in this commit.

* Adding snapshot tests for the deposits information block.

* Adding selector tests

* added op chaining, in-case a group value is undefined

* Removing the single deposit snapshot test,
since everything needed is handled when rendering multiple currencies.

* Extracting components into separate files

DepositsInformationBlock and DepositsInformationLoading are now separate files to simplify the focus in the main deposits information block file.

* Testing for the instant deposit button for specificity.

* Changing the missing information test to add more specificity.

* Using universal breakpoints for the overview page

The breakpoint function, along with a standardized width, is much better than a custom media query.

* Adding the missing tests coverage to `deposits/utils`

* Updated types

* Removing un-neccessary module declarations

* Improving the gridicons declaration by adding particular types/params

* Adding changelog entry

* Limiting the overview page to 680 pixels wide

Co-authored-by: Radoslav Georgiev <rageorgiev@gmail.com>
Co-authored-by: Naman Malhotra <naman03malhotra@gmail.com>